refactor: migrate agent polling out of app context
