+++ webui/react/src/pages/Cluster.tsx
@@ -1,6 +1,6 @@
 import { Radio } from 'antd';
 import { RadioChangeEvent } from 'antd/lib/radio';
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import Grid, { GridMode } from 'components/Grid';
 import Icon from 'components/Icon';
@@ -13,7 +13,7 @@ import Section from 'components/Section';
 import SlotAllocationBar from 'components/SlotAllocationBar';
 import { Indicator } from 'components/Spinner';
 import { defaultRowClassName, getPaginationConfig, isAlternativeAction } from 'components/Table';
-import Agents from 'contexts/Agents';
+import Agents, { useFetchAgents } from 'contexts/Agents';
 import ClusterOverview, { agentsToOverview } from 'contexts/ClusterOverview';
 import usePolling from 'hooks/usePolling';
 import useStorage from 'hooks/useStorage';
@@ -41,13 +41,21 @@ const Cluster: React.FC = () => {
   const [ rpDetail, setRpDetail ] = useState<ResourcePool>();
   const [ selectedView, setSelectedView ] = useState<View>(initView);
   const [ resourcePools, setResourcePools ] = useState<ResourcePool[]>([]);
+  const [ canceler ] = useState(new AbortController());
 
-  const pollResourcePools = useCallback(async () => {
+  const fetchAgents = useFetchAgents(canceler);
+
+  const fetchResourcePools = useCallback(async () => {
     const resourcePools = await getResourcePools({});
     setResourcePools(resourcePools);
   }, []);
 
-  usePolling(pollResourcePools, { delay: 10000 });
+  const fetchAll = useCallback(() => {
+    fetchAgents();
+    fetchResourcePools();
+  }, [ fetchAgents, fetchResourcePools ]);
+
+  usePolling(fetchAll, { delay: 10000 });
 
   const cpuContainers = useMemo(() => {
     const tally = {
@@ -116,6 +124,10 @@ const Cluster: React.FC = () => {
     };
     return { onAuxClick: handleClick, onClick: handleClick };
   }, []);
+
+  useEffect(() => {
+    return () => canceler.abort();
+  }, [ canceler ]);
 
   const viewOptions = (
     <Radio.Group value={selectedView} onChange={onChange}>
